requirements.txt
streamlit
openai>=1.52.0
python-dotenv
apscheduler
pytz